Remove horizontal scrollbar in data table
Set columns to fixed width and text-overflow: ellipsis.
Added titles to the columns which are most likely to overflow,
so the user can still access the full contents of the cell if
the cell overflows.

Prevents a horizontal scrollbar from showing up by ensuring
the table content doesn't increase its width.